requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
Faker==15.3.4
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pydantic==2.3.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
urllib3==1.26.14
